Bring back deprecated notices.
