+++ www/src/components/thumbprint-components/component-header/component-package-table.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import PackageTable from '../package-table';
+import PackageTable from '../../package-table';
 
 /**
  * Returns an JavaScript import statement for components (and subcomponents) within a package.
@@ -13,7 +13,6 @@ const ComponentPackageTable = ({ version, name, homepage, components, deprecated
         version={version}
         packageName={name}
         sourceDirectory={homepage}
-        deprecated={deprecated}
         importStatement={
             components
                 ? getImportStatement({
@@ -30,12 +29,10 @@ ComponentPackageTable.propTypes = {
     name: PropTypes.string.isRequired,
     homepage: PropTypes.string.isRequired,
     components: PropTypes.arrayOf(PropTypes.shape({})),
-    deprecated: PropTypes.oneOfType(PropTypes.bool, PropTypes.string),
 };
 
 ComponentPackageTable.defaultProps = {
     components: undefined,
-    deprecated: false,
 };
 
 export default ComponentPackageTable;
