+++ www/src/pages/components/modal-standard/react/index.mdx
@@ -3,10 +3,19 @@ title: Modal Standard
 description: A basic and commonly used modal.
 ---
 
-import { graphql } from 'gatsby';
-import { ComponentHeader, ComponentFooter } from 'components/thumbprint-components';
+import { graphql, Link } from 'gatsby';
+import {
+    ComponentHeader,
+    ComponentFooter,
+    DeprecatedComponentAlert,
+} from 'components/thumbprint-components';
 
 <ComponentHeader data={props.data} />
+
+<DeprecatedComponentAlert>
+    This package was deprecated in favor of{' '}
+    <Link to="/components/modal-curtain/react/">ModalCurtain</Link>.
+</DeprecatedComponentAlert>
 
 ## Basic modal
